Add Education section
Add the education section. This exposed chances to refactor out list
styling and to reuse styling from job (which required small rewrite of
job.pug)

As well, a common font-size of 14px was put on the Miscellaneous section
but since headers were defined in em's, this broke consistent header
sizing. To fix this, headers were redefined in terms of pixels.